refactor: move all identity related functions from PlatformRepo to IdentityRepository
