+++ wallet/src/de/schildbach/wallet/di/AppModule.kt
@@ -38,6 +38,7 @@ import de.schildbach.wallet.service.*
 import de.schildbach.wallet.service.AndroidActionsService
 import de.schildbach.wallet.service.AppRestartService
 import de.schildbach.wallet.service.RestartService
+import de.schildbach.wallet.service.platform.IdentityRepository
 import de.schildbach.wallet.ui.dashpay.PlatformRepo
 import de.schildbach.wallet.ui.more.tools.ZenLedgerApi
 import de.schildbach.wallet.ui.more.tools.ZenLedgerClient
@@ -112,10 +113,11 @@ abstract class AppModule {
             identityConfig: BlockchainIdentityConfig,
             coinJoinConfig: CoinJoinConfig,
             coinJoinService: CoinJoinService,
+            identityRepository: IdentityRepository,
             platformRepo: PlatformRepo,
             transactionMetadataProvider: TransactionMetadataProvider
         ): SendPaymentService {
-            val realService = SendCoinsTaskRunner(walletData, walletApplication, securityFunctions, packageInfoProvider, analyticsService, identityConfig, coinJoinConfig, coinJoinService, platformRepo, transactionMetadataProvider)
+            val realService = SendCoinsTaskRunner(walletData, walletApplication, securityFunctions, packageInfoProvider, analyticsService, identityConfig, coinJoinConfig, coinJoinService, identityRepository, platformRepo, transactionMetadataProvider)
 
             return if (BuildConfig.FLAVOR.lowercase() == "prod") {
                 realService
