+++ wallet/src/de/schildbach/wallet/payments/SendCoinsTaskRunner.kt
@@ -31,6 +31,7 @@ import de.schildbach.wallet.service.CoinJoinMode
 import de.schildbach.wallet.service.CoinJoinService
 import de.schildbach.wallet.service.MixingStatus
 import de.schildbach.wallet.service.PackageInfoProvider
+import de.schildbach.wallet.service.platform.IdentityRepository
 import de.schildbach.wallet.ui.dashpay.PlatformRepo
 import de.schildbach.wallet.util.AnrException
 import kotlinx.coroutines.*
@@ -80,6 +81,7 @@ class SendCoinsTaskRunner @Inject constructor(
     private val identityConfig: BlockchainIdentityConfig,
     coinJoinConfig: CoinJoinConfig,
     coinJoinService: CoinJoinService,
+    private val identityRepository: IdentityRepository,
     private val platformRepo: PlatformRepo,
     private val metadataProvider: TransactionMetadataProvider
 ) : SendPaymentService {
@@ -668,7 +670,7 @@ class SendCoinsTaskRunner @Inject constructor(
                 it.value.value
             }
             val isSentToContact = try {
-                platformRepo.blockchainIdentity.getContactForTransaction(transaction) != null
+                identityRepository.blockchainIdentity?.getContactForTransaction(transaction) != null
             } catch (e: Exception) {
                 false
             }
