+++ wallet/src/de/schildbach/wallet/service/BlockchainServiceImpl.kt
@@ -58,6 +58,7 @@ import de.schildbach.wallet.data.AddressBookProvider
 import de.schildbach.wallet.database.dao.BlockchainStateDao
 import de.schildbach.wallet.database.dao.ExchangeRatesDao
 import de.schildbach.wallet.service.extensions.registerCrowdNodeConfirmedAddressFilter
+import de.schildbach.wallet.service.platform.IdentityRepository
 import de.schildbach.wallet.service.platform.PlatformSyncService
 import de.schildbach.wallet.transactions.WalletObserver.Companion.CONFIRMED_DEPTH
 import de.schildbach.wallet.service.platform.TopUpRepository
@@ -246,8 +247,8 @@ class BlockchainServiceImpl : LifecycleService(), BlockchainService {
     @Inject lateinit var transactionMetadataProvider: TransactionMetadataProvider
 
     @Inject lateinit var platformSyncService: PlatformSyncService
-
-    @Inject lateinit var  platformRepo: PlatformRepo
+    @Inject lateinit var identityRepo: IdentityRepository
+    @Inject lateinit var platformRepo: PlatformRepo
     @Inject lateinit var topUpRepository: TopUpRepository
 
     @Inject lateinit var  packageInfoProvider: PackageInfoProvider
@@ -807,7 +808,7 @@ class BlockchainServiceImpl : LifecycleService(), BlockchainService {
                 updateBlockchainState()
                 if (Constants.SUPPORTS_PLATFORM) {
                     serviceScope.launch {
-                        platformRepo.updateFrequentContacts()
+                        identityRepo.updateFrequentContacts()
                     }
                 }
             }
@@ -2225,7 +2226,7 @@ class BlockchainServiceImpl : LifecycleService(), BlockchainService {
                 val txsToProcess = synchronized(pendingContactPaymentTxs) {
                     pendingContactPaymentTxs.toList().also { pendingContactPaymentTxs.clear() }
                 }
-                txsToProcess.forEach { platformRepo.updateFrequentContacts(it) }
+                txsToProcess.forEach { identityRepo.updateFrequentContacts(it) }
             }
         }
     }
