+++ wallet/src/de/schildbach/wallet/service/platform/TopUpRepository.kt
@@ -159,6 +159,7 @@ interface TopUpRepository {
 class TopUpRepositoryImpl @Inject constructor(
     private val walletApplication: WalletApplication,
     private val walletDataProvider: WalletDataProvider,
+    private val identityRepository: IdentityRepository,
     private val platformRepo: PlatformRepo,
     private val topUpsDao: TopUpsDao,
     private val dashPayProfileDao: DashPayProfileDao,
@@ -381,7 +382,7 @@ class TopUpRepositoryImpl @Inject constructor(
     private suspend fun addTopUp(txId: Sha256Hash): TopUp {
         val topUp = TopUp(
             txId,
-            platformRepo.blockchainIdentity.uniqueIdString
+            identityRepository.blockchainIdentity!!.uniqueIdString
         )
         topUpsDao.insert(topUp)
         return topUp
@@ -416,7 +417,7 @@ class TopUpRepositoryImpl @Inject constructor(
         }
         log.info("topup tx sent: {}; tx = {}", status, topUpTx.txId)
         try {
-            platformRepo.blockchainIdentity.topUp(
+            identityRepository.blockchainIdentity!!.topUp(
                 topUpTx,
                 aesKeyParameter,
                 useISLock = true,
@@ -538,7 +539,7 @@ class TopUpRepositoryImpl @Inject constructor(
             authExtension.topupFundingTransactions.forEach { assetLockTx ->
                 val topUp = topUpsDao.getByTxId(assetLockTx.txId)
                 if (topUp == null || topUp.notUsed()) {
-                    val identity = topUp?.toUserId ?: platformRepo.blockchainIdentity.uniqueIdentifier.toString()
+                    val identity = topUp?.toUserId ?: identityRepository.blockchainIdentity!!.uniqueIdentifier.toString()
                     if (topUp == null) {
                         topUpsDao.insert(TopUp(assetLockTx.txId, identity))
                     }
@@ -671,7 +672,7 @@ class TopUpRepositoryImpl @Inject constructor(
                         fundingTxes.remove(invitation.txid)
                     } else {
                         // TODO: should we fix the link now or let the user do it
-                        val dashPayProfile = platformRepo.getLocalUserProfile()
+                        val dashPayProfile = identityRepository.getLocalUserProfile()
                         val assetLockTx = fundingTxes[invitation.txid]
                         if (assetLockTx != null) {
                             val appsFlyerLink = createAppsFlyerLink(dashPayProfile!!, assetLockTx, encryptionKey)
@@ -736,7 +737,7 @@ class TopUpRepositoryImpl @Inject constructor(
     override fun handleSentAssetLockTransaction(cftx: AssetLockTransaction, blockTimestamp: Long) {
         val extension = authExtension
 
-        if (platformRepo.hasBlockchainIdentity) {
+        if (identityRepository.hasBlockchainIdentity) {
             workerScope.launch(Dispatchers.IO) {
                 // Context.getOrCreate(platform.params)
                 val inviteKey = extension.invitationFundingKeyChain.findKeyFromPubHash(cftx.assetLockPublicKeyId.bytes)
