+++ wallet/src/de/schildbach/wallet/ui/dashpay/PlatformRepo.kt
@@ -65,7 +65,6 @@ import org.dash.wallet.common.services.analytics.AnalyticsService
 import org.dash.wallet.common.services.analytics.AnalyticsTimer
 import org.dashj.platform.dapiclient.MaxRetriesReachedException
 import org.dashj.platform.dapiclient.NoAvailableAddressesForRetryException
-import org.dashj.platform.dapiclient.model.GrpcExceptionInfo
 import org.dashj.platform.dashpay.*
 import org.dashj.platform.dpp.document.Document
 import org.dashj.platform.dpp.errors.concensus.basic.identity.InvalidInstantAssetLockProofException
@@ -74,6 +73,7 @@ import org.dashj.platform.dpp.identity.Identity
 import org.dashj.platform.dpp.voting.Contenders
 import org.dashj.platform.sdk.platform.DomainDocument
 import org.dashj.platform.sdk.platform.Names
+import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.util.*
 import javax.inject.Inject
@@ -81,11 +81,9 @@ import javax.inject.Singleton
 import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
 
-@OptIn(ExperimentalCoroutinesApi::class)
 @Singleton
 class PlatformRepo @Inject constructor(
     val walletApplication: WalletApplication,
-    val blockchainIdentityDataStorage: BlockchainIdentityConfig,
     val appDatabase: AppDatabase,
     val platform: PlatformService,
     val coinJoinConfig: CoinJoinConfig,
@@ -106,60 +104,8 @@ class PlatformRepo @Inject constructor(
 
     private val onSeriousErrorListeneners = arrayListOf<SeriousErrorListener>()
 
-    lateinit var blockchainIdentity: BlockchainIdentity
-        private set
-
-    val hasBlockchainIdentity: Boolean
-        get() = this::blockchainIdentity.isInitialized
-
-    suspend fun hasIdentity(): Boolean = this::blockchainIdentity.isInitialized ||
-            blockchainIdentityDataStorage.get(BlockchainIdentityConfig.IDENTITY_ID) != null
-
-    suspend fun hasUsername(): Boolean = (this::blockchainIdentity.isInitialized && blockchainIdentity.currentUsername != null) ||
-            blockchainIdentityDataStorage.get(BlockchainIdentityConfig.USERNAME) != null
-
-    @Throws(IllegalStateException::class)
-    suspend fun getIdentity(): String {
-        return if (this::blockchainIdentity.isInitialized) {
-            blockchainIdentity.uniqueIdString
-        } else {
-            blockchainIdentityDataStorage.get(BlockchainIdentityConfig.IDENTITY_ID)
-                ?: throw IllegalStateException("IdentityId not found")
-        }
-    }
-
-    suspend fun getUsername(): String? {
-        return if (this::blockchainIdentity.isInitialized) {
-            blockchainIdentity.currentUsername
-        } else {
-            blockchainIdentityDataStorage.get(BlockchainIdentityConfig.USERNAME)
-        }
-    }
-
-    suspend fun getActiveUsername(): String? {
-        return if (this::blockchainIdentity.isInitialized) {
-            blockchainIdentity.currentUsername
-        } else {
-            val username = blockchainIdentityDataStorage.get(BlockchainIdentityConfig.USERNAME)
-            val creationState = blockchainIdentityDataStorage.get(BlockchainIdentityConfig.CREATION_STATE)
-            if (creationState == IdentityCreationState.VOTING.name) {
-                val usernameSecondary = blockchainIdentityDataStorage.get(BlockchainIdentityConfig.USERNAME_SECONDARY)
-                if (usernameSecondary != null &&
-                    blockchainIdentityDataStorage.get(BlockchainIdentityConfig.USERNAME_SECONDARY_REGISTRATION_STATUS) == UsernameStatus.CONFIRMED.name) {
-                    usernameSecondary
-                } else {
-                    username
-                }
-            } else {
-                username
-            }
-        }
-    }
-
     val authenticationGroupExtension: AuthenticationGroupExtension?
         get() = walletApplication.authenticationGroupExtension
-    //getKeyChainExtension(AuthenticationGroupExtension.EXTENSION_ID) as? AuthenticationGroupExtension
-
 
     private val dashPayProfileDao = appDatabase.dashPayProfileDao()
     private val dashPayContactRequestDao = appDatabase.dashPayContactRequestDao()
@@ -183,13 +129,6 @@ class PlatformRepo @Inject constructor(
     init {
         backgroundThread.start()
         backgroundHandler = Handler(backgroundThread.looper)
-    }
-
-    suspend fun init() {
-        blockchainIdentityDataStorage.load()?.let {
-            blockchainIdentity = initBlockchainIdentity(it, walletApplication.wallet!!)
-            initializeStateRepository()
-        }
     }
 
     fun getWalletEncryptionKey(): KeyParameter? {
@@ -230,50 +169,12 @@ class PlatformRepo @Inject constructor(
         }
     }
 
-    /**
-     * This method looks at all items in the database tables
-     * that have existing identites and saves them for future use.
-     *
-     * Sometimes Platform Nodes return IdentityNotFound Errors and
-     * this list is used to determine if that node should be banned
-     */
-    private suspend fun initializeStateRepository() {
-        // load our id
-
-        if (this::blockchainIdentity.isInitialized && blockchainIdentity.isRegistered()) {
-            val identityId = blockchainIdentity.uniqueIdString
-            platform.stateRepository.addValidIdentity(Identifier.from(identityId))
-
-            // load all id's of users who have sent us a contact request
-            dashPayContactRequestDao.loadFromOthers(identityId).forEach {
-                platform.stateRepository.addValidIdentity(it.userIdentifier)
-            }
-
-            // load all id's of users for whom we have profiles
-            dashPayProfileDao.loadAll().forEach {
-                platform.stateRepository.addValidIdentity(it.userIdentifier)
-            }
-
-            platform.stateRepository.storeIdentity(blockchainIdentity.identity!!)
-        }
-    }
-
     fun getUsername(username: String): Resource<Document> {
         return try {
             val nameDocument = platform.names.get(Names.normalizeString(username))
             Resource.success(nameDocument)
         } catch (e: Exception) {
             Resource.error(e.localizedMessage!!, null)
-        }
-    }
-
-    @Throws(Exception::class)
-    suspend fun getUser(username: String): List<UsernameSearchResult> {
-        return try {
-            searchUsernames(username, true)
-        } catch (e: Exception) {
-            formatExceptionMessage("get single user failure", e)
-            throw e
         }
     }
 
@@ -288,327 +189,7 @@ class PlatformRepo @Inject constructor(
         }
     }
 
-    /**
-     * gets all the name documents for usernames starting with text
-     *
-     * @param text The beginning of a username to search for
-     * @return
-     */
-
-    @Throws(Exception::class)
-    suspend fun searchUsernames(text: String, onlyExactUsername: Boolean = false, limit: Int = -1): List<UsernameSearchResult> {
-        return withContext(Dispatchers.IO) {
-            val userIdString = blockchainIdentity.uniqueIdString
-            val userId = blockchainIdentity.uniqueIdentifier
-
-            // Names.search does support retrieving 100 names at a time if retrieveAll = false
-            //TODO: Maybe add pagination later? Is very unlikely that a user will scroll past 100 search results
-            // Sometimes when onlyExactUsername = true, an exception is thrown here and that results in a crash
-            // it is not clear why a search for an existing username results in a failure to find it again.
-            val nameDocuments = if (!onlyExactUsername) {
-                platform.names.search(text, Names.DEFAULT_PARENT_DOMAIN, retrieveAll = false, limit = limit)
-            } else {
-                val nameDocument = platform.names.get(text, Names.DEFAULT_PARENT_DOMAIN)
-                if (nameDocument != null) {
-                    listOf(nameDocument)
-                } else {
-                    listOf()
-                }
-            }
-            // determine if multiple names belong to the same identity. If so, don't show any non-contested names
-            val identifierDocumentMap = hashMapOf<Identifier, ArrayList<DomainDocument>>()
-            nameDocuments.forEach { document ->
-                val domainDocument = DomainDocument(document)
-                val identifier = getIdentityForName(domainDocument)
-                if (identifierDocumentMap.contains(identifier)) {
-                    identifierDocumentMap[identifier]?.add(domainDocument)
-                } else {
-                    val newList = arrayListOf(domainDocument)
-                    identifierDocumentMap[identifier] = newList
-                }
-            }
-
-            val userIds = if (onlyExactUsername) {
-                val result = mutableListOf<Identifier>()
-                val exactNameDoc = try {
-                    DomainDocument(nameDocuments.first { text == it.data["normalizedLabel"] })
-                } catch (e: NoSuchElementException) {
-                    null
-                }
-                if (exactNameDoc != null) {
-                    result.add(getIdentityForName(exactNameDoc))
-                }
-                result
-            } else {
-                identifierDocumentMap.keys
-            }.toSet().toList()
-
-            val profileById: Map<Identifier, Document> = if (userIds.isNotEmpty()) {
-                val profileDocuments = platform.profiles.getList(userIds)
-                profileDocuments.associateBy({ it.ownerId }, { it })
-            } else {
-                log.warn("search usernames: userIdList is empty, though nameDocuments has ${nameDocuments.size} items")
-                mapOf()
-            }
-
-            // remove non-contested names if there are contested names for the same identity
-            val filteredNameDocuments = arrayListOf<DomainDocument>()
-            identifierDocumentMap.forEach { (identifier, documents) ->
-                if (documents.size == 1) {
-                    filteredNameDocuments.addAll(documents)
-                } else {
-                    val hasContestedNames = documents.any { Names.isUsernameContestable(it.normalizedLabel) }
-                    documents.forEach { document ->
-                        if (Names.isUsernameContestable(document.normalizedLabel)) {
-                            filteredNameDocuments.add(document)
-                        } else if (!hasContestedNames) {
-                            filteredNameDocuments.add(document)
-                        }
-                    }
-                }
-            }
-
-            val toContactDocuments = dashPayContactRequestDao.loadToOthers(userIdString)
-
-            // Get all contact requests where toUserId == userId
-            val fromContactDocuments = dashPayContactRequestDao.loadFromOthers(userIdString)
-
-            val usernameSearchResults = ArrayList<UsernameSearchResult>()
-
-            for (nameDoc in filteredNameDocuments) {
-                if (nameDoc.dashAliasIdentityId != null) {
-                    continue // skip aliases
-                }
-
-                //Remove own user document from result
-                val nameDocIdentityId = getIdentityForName(nameDoc)
-                if (nameDocIdentityId == userId) {
-                    continue
-                }
-                var toContact: DashPayContactRequest? = null
-                var fromContact: DashPayContactRequest? = null
-
-                // Determine if any of our contacts match the current name's identity
-                if (toContactDocuments.isNotEmpty()) {
-                    toContact = toContactDocuments.find { contact ->
-                        contact.toUserIdentifier == nameDocIdentityId
-                    }
-                }
-
-                // Determine if our identity is someone else's contact
-                if (fromContactDocuments.isNotEmpty()) {
-                    fromContact = fromContactDocuments.find { contact ->
-                        contact.userIdentifier == nameDocIdentityId
-                    }
-                }
-
-                val username = nameDoc.label
-                val profileDoc = profileById[nameDocIdentityId]
-
-                val dashPayProfile = if (profileDoc != null)
-                    DashPayProfile.fromDocument(profileDoc, username)!!
-                else DashPayProfile(nameDocIdentityId.toString(), username)
-
-                usernameSearchResults.add(UsernameSearchResult(username,
-                    dashPayProfile, toContact, fromContact))
-            }
-
-            // TODO: this is only needed when Proofs don't sort results
-            // This was added in v0.20
-            usernameSearchResults.sortBy { Names.normalizeString(it.username) }
-
-            return@withContext usernameSearchResults
-        }
-    }
-
-    /**
-     * search the contacts
-     *
-     * @param text the text to find in usernames and displayNames.  if blank, all contacts are returned
-     * @param orderBy the field that is used to sort the list of matching entries in ascending order
-     * @return
-     */
-    suspend fun searchContacts(text: String, orderBy: UsernameSortOrderBy, includeSentPending: Boolean = false): Resource<List<UsernameSearchResult>> {
-        if (!hasIdentity()) {
-            return Resource.success(emptyList())
-        }
-
-        return try {
-            val userIdList = HashSet<String>()
-
-            val userId = getIdentity()
-
-            val toContactDocuments = dashPayContactRequestDao.loadToOthers(userId)
-            val toContactMap = HashMap<String, DashPayContactRequest>()
-            toContactDocuments.forEach {
-                userIdList.add(it.toUserId)
-                toContactMap[it.toUserId] = it
-            }
-            // Get all contact requests where toUserId == userId, the users who have added me
-            val fromContactDocuments = dashPayContactRequestDao.loadFromOthers(userId)
-            val fromContactMap = HashMap<String, DashPayContactRequest>()
-            fromContactDocuments.forEach {
-                userIdList.add(it.userId)
-
-                // It is possible for a contact to send multiple requests that differ by account
-                // or by version.  Currently we will ignore all but the first based on the timestamp
-                // TODO: choose the contactRequest based on the ContactInfo.accountRef value
-                // for this contact
-                if (!fromContactMap.containsKey(it.userId)) {
-                    fromContactMap[it.userId] = it
-                } else {
-                    val previous = fromContactMap[it.userId]!!
-                    if (previous.timestamp > it.timestamp) {
-                        fromContactMap[it.userId] = it
-                    }
-                }
-            }
-
-            val profiles = HashMap<String, DashPayProfile?>(userIdList.size)
-            for (user in userIdList) {
-                val profile = dashPayProfileDao.loadByUserId(user)
-                profiles[user] = profile
-            }
-
-            val usernameSearchResults = getFromProfiles(profiles, text.lowercase(), toContactMap, fromContactMap, includeSentPending)
-            usernameSearchResults.orderBy(orderBy)
-
-            Resource.success(usernameSearchResults)
-        } catch (e: Exception) {
-            Resource.error(formatExceptionMessage("search contact request", e), null)
-        }
-    }
-
-    fun observeContacts(text: String, orderBy: UsernameSortOrderBy, includeSentPending: Boolean = false): Flow<List<UsernameSearchResult>> {
-        return blockchainIdentityDataStorage.observe()
-            .filterNotNull()
-            .filter { it.hasUsername }
-            .flatMapLatest { identityData ->
-                val userId = identityData.userId!!
-
-                // Combine the two contact request flows
-                combine(
-                    dashPayContactRequestDao.observeToOthers(userId),
-                    dashPayContactRequestDao.observeFromOthers(userId)
-                ) { toContacts, fromContacts ->
-                    val userIdList = HashSet<String>()
-
-                    val toContactMap = HashMap<String, DashPayContactRequest>()
-                    toContacts.forEach {
-                        userIdList.add(it.toUserId)
-                        toContactMap[it.toUserId] = it
-                    }
-
-                    val fromContactMap = HashMap<String, DashPayContactRequest>()
-                    fromContacts.forEach {
-                        userIdList.add(it.userId)
-                        if (!fromContactMap.containsKey(it.userId)) {
-                            fromContactMap[it.userId] = it
-                        } else {
-                            val previous = fromContactMap[it.userId]!!
-                            if (previous.timestamp > it.timestamp) {
-                                fromContactMap[it.userId] = it
-                            }
-                        }
-                    }
-
-                    Triple(userIdList, toContactMap, fromContactMap)
-                }.flatMapLatest { (userIdList, toContactMap, fromContactMap) ->
-                    dashPayProfileDao.observeByUserIds(userIdList.toList()).map { list ->
-                        val profiles = list.associateBy { it.userId }
-                        val usernameSearchResults = getFromProfiles(profiles, text.lowercase(), toContactMap, fromContactMap, includeSentPending)
-                        usernameSearchResults.orderBy(orderBy)
-                        usernameSearchResults
-                    }
-                }
-            }
-            .distinctUntilChanged()
-    }
-
-    suspend fun updateFrequentContacts(newTx: Transaction) {
-        // since we are accessing the blockchainIdentity object, we better check that it is valid
-        // previously, we were using hasUsername() which can return true during a wallet reset
-        if (hasBlockchainIdentity && blockchainIdentity.getContactForTransaction(newTx) != null) {
-            updateFrequentContacts()
-        }
-    }
-
-    suspend fun updateFrequentContacts() {
-        if (hasIdentity()) {
-            val contactRequests = searchContacts("", UsernameSortOrderBy.DATE_ADDED)
-            val frequentContacts = when (contactRequests.status) {
-                Status.SUCCESS -> {
-                    if (!hasBlockchainIdentity) {
-                        return
-                    }
-
-                    val threeMonthsAgo = Date().time - TIMESPAN
-
-                    val results =
-                        getTopContacts(contactRequests.data!!, listOf(), blockchainIdentity, threeMonthsAgo, true)
-
-                    if (results.size < TOP_CONTACT_COUNT) {
-                        val moreResults =
-                            getTopContacts(contactRequests.data, results, blockchainIdentity, threeMonthsAgo, false)
-                        results.addAll(moreResults)
-                    }
-
-                    results
-                }
-
-                else -> listOf<UsernameSearchResult>()
-            }
-            dashPayConfig.set(DashPayConfig.FREQUENT_CONTACTS, frequentContacts.map { it.getIdentity() }.toSet())
-        }
-    }
-
-    private fun getTopContacts(items: List<UsernameSearchResult>,
-                               ignore: List<UsernameSearchResult>,
-                               blockchainIdentity: BlockchainIdentity,
-                               threeMonthsAgo: Long,
-                               sent: Boolean
-    ): ArrayList<UsernameSearchResult> {
-        val wholeWatch = Stopwatch.createStarted()
-        val results = arrayListOf<UsernameSearchResult>()
-        val contactScores = hashMapOf<String, Int>()
-        val contactIds = arrayListOf<String>()
-        // only include fully established contacts
-        val contacts = items.filter { it.requestSent && it.requestReceived }
-
-        contacts.forEach {
-            val watch = Stopwatch.createStarted()
-            val transactions = blockchainIdentity.getContactTransactions(it.fromContactRequest!!.userIdentifier, it.fromContactRequest!!.accountReference)
-            var count = 0
-
-            for (tx in transactions) {
-                val txValue = tx.getValue(walletApplication.wallet)
-                if ((sent && txValue.isNegative) || (!sent && txValue.isPositive)) {
-                    if (tx.updateTime.time > threeMonthsAgo) {
-                        count++
-                    }
-                }
-            }
-            contactScores[it.fromContactRequest!!.userId] = count
-            contactIds.add(it.fromContactRequest!!.userId)
-        }
-
-        // determine users with top TOP_CONTACT_COUNT non-zero scores
-        // if ignore has some items, then find TOP_CONTACT_COUNT - ignore.size
-        contactIds.sortByDescending { contactScores[it] }
-        var count = 0
-        for (id in contactIds) {
-            if (contactScores[id] != 0 && ignore.find { it.fromContactRequest!!.userId == id } == null) {
-                results.add(items.find { it.fromContactRequest!!.userId == id }!!)
-                count++
-                if (count == TOP_CONTACT_COUNT - ignore.size)
-                    break
-            }
-        }
-        log.info("frequent processing: {}", wholeWatch)
-        return results
-    }
-
-    private fun getFromProfiles(
+    fun getFromProfiles(
         profiles: Map<String, DashPayProfile?>,
         searchText: String,
         toContactMap: Map<String, DashPayContactRequest>,
@@ -652,6 +233,10 @@ class PlatformRepo @Inject constructor(
     }
 
     fun formatExceptionMessage(description: String, e: Exception): String {
+        return formatExceptionMessage(description, e, log)
+    }
+
+    fun formatExceptionMessage(description: String, e: Exception, log: Logger): String {
         var msg = if (e.localizedMessage != null) {
             e.localizedMessage
         } else {
@@ -664,49 +249,6 @@ class PlatformRepo @Inject constructor(
         return msg
     }
 
-    /**
-     * returns true if:
-     *  1. Invites have been not been sent previously
-     *  2. Identity Creation is not in progress
-     *  3. The balance is high enough
-     */
-
-    suspend fun shouldShowAlert(): Boolean {
-        val hasSentInvites = invitationsDao.count() > 0
-        val blockchainIdentityData = blockchainIdentityDataStorage.load()
-        val noIdentityCreatedOrInProgress = (blockchainIdentityData == null) || blockchainIdentityData.creationState == IdentityCreationState.NONE
-        val canAffordIdentityCreation = walletApplication.canAffordIdentityCreation()
-        return !noIdentityCreatedOrInProgress && (canAffordIdentityCreation || hasSentInvites)
-    }
-
-
-    suspend fun getNotificationCount(date: Long): Int {
-        var count = 0
-        if (!isUsernameRegistered()) {
-            return 0
-        }
-
-        if (Constants.SUPPORTS_INVITES) {
-            if (shouldShowAlert()) {
-                val alert = userAlertDao.load(date)
-                if (alert != null) {
-                    count++
-                }
-            }
-        }
-
-        val results = searchContacts("", UsernameSortOrderBy.DATE_ADDED)
-        if (results.status == Status.SUCCESS) {
-            val list = results.data ?: return 0
-            list.forEach { if (it.date >= date) ++count }
-            log.info("New contacts at ${Date(date)} = $count - getNotificationCount")
-        }
-        return count
-    }
-
-    private fun isUsernameRegistered(): Boolean {
-        return this::blockchainIdentity.isInitialized
-    }
 
     //
     // Step 1 is to upgrade the wallet to support authentication keys
@@ -790,153 +332,10 @@ class PlatformRepo @Inject constructor(
         }
     }
 
-    suspend fun loadBlockchainIdentityBaseData(): BlockchainIdentityBaseData? {
-        return blockchainIdentityDataStorage.loadBase()
-    }
 
-    suspend fun loadBlockchainIdentityData(): BlockchainIdentityData? {
-        return blockchainIdentityDataStorage.load()
-    }
 
-    fun initBlockchainIdentity(blockchainIdentityData: BlockchainIdentityData, wallet: Wallet): BlockchainIdentity {
-        // previously, we would look up the asset lock transaction, but we don't need to do that
-        val watch = Stopwatch.createStarted()
-        log.info("loading BlockchainIdentity: starting...")
-        val authExt = authenticationGroupExtension
-            ?: throw IllegalStateException("AuthenticationGroupExtension is not initialised")
-        val blockchainIdentity = BlockchainIdentity(platform.platform, 0, wallet, authExt)
-        log.info("loading BlockchainIdentity: {}", watch)
-        if (blockchainIdentityData.creationState >= IdentityCreationState.IDENTITY_REGISTERED) {
-            blockchainIdentity.apply {
-                blockchainIdentityData.userId?.let {
-                    uniqueId = Sha256Hash.wrap(Base58.decode(it))
-                }
-                identity = blockchainIdentityData.identity
-            }
-            blockchainIdentity.updateIdentity()
-            log.info("loading identity ${blockchainIdentityData.userId} == ${if (this::blockchainIdentity.isInitialized) blockchainIdentity.uniqueIdString else null}: {}", watch)
-        } else {
-            log.info("loading identity: {}", watch)
-            return blockchainIdentity
-        }
-
-        // TODO: needs to check against Platform to see if values exist.  Check after
-        // Syncing complete
-        log.info("loading identity ${blockchainIdentityData.userId} == ${if (this::blockchainIdentity.isInitialized) blockchainIdentity.uniqueIdString else null}: {}", watch)
-        return blockchainIdentity.apply {
-            primaryUsername = blockchainIdentityData.username
-            secondaryUsername = blockchainIdentityData.usernameSecondary
-            blockchainIdentityData.username?.let {
-                addUsername(it)
-            }
-            blockchainIdentityData.usernameSecondary?.let {
-                addUsername(it)
-            }
-
-            registrationStatus = blockchainIdentityData.registrationStatus ?: IdentityStatus.NOT_REGISTERED
-            // usernameStatus, usernameSalts are not set if preorder hasn't started
-            if (blockchainIdentityData.creationState >= IdentityCreationState.PREORDER_REGISTERING) {
-                val usernameStatus = UsernameInfo(
-                    blockchainIdentityData.preorderSalt,
-                    blockchainIdentityData.usernameStatus ?: UsernameStatus.NOT_PRESENT,
-                    currentUsername,
-                    blockchainIdentityData.usernameRequested,
-                    blockchainIdentityData.votingPeriodStart
-                )
-                currentUsername ?.let {
-                    usernameStatuses[it] = usernameStatus
-                }
-            }
-
-            if (blockchainIdentityData.creationState >= IdentityCreationState.PREORDER_SECONDARY_REGISTERING) {
-                val usernameStatus = UsernameInfo(
-                    blockchainIdentityData.preorderSaltSecondary,
-                    blockchainIdentityData.usernameSecondaryStatus ?: UsernameStatus.NOT_PRESENT,
-                    blockchainIdentityData.usernameSecondary,
-                    null,
-                    null
-                )
-                secondaryUsername ?.let {
-                    usernameStatuses[it] = usernameStatus
-                }
-            }
-
-            creditBalance = blockchainIdentityData.creditBalance ?: Coin.ZERO
-            log.info("loading identity: {}", watch)
-        }
-    }
-
-    suspend fun updateBlockchainIdentityData() {
-        blockchainIdentityDataStorage.load()?.let {
-            updateBlockchainIdentityData(it, blockchainIdentity)
-        }
-    }
-
-    suspend fun updateBlockchainIdentityData(blockchainIdentityData: BlockchainIdentityData, blockchainIdentity: BlockchainIdentity) {
-        blockchainIdentityData.apply {
-            creditFundingTxId = blockchainIdentity.assetLockTransaction?.txId
-            userId = if (blockchainIdentity.registrationStatus == IdentityStatus.REGISTERED)
-                blockchainIdentity.uniqueIdString
-            else null
-            identity = blockchainIdentity.identity
-            registrationStatus = blockchainIdentity.registrationStatus
-            if (blockchainIdentity.currentUsername != null) {
-                username = blockchainIdentity.primaryUsername
-                if (blockchainIdentity.registrationStatus == IdentityStatus.REGISTERED) {
-                    preorderSalt = blockchainIdentity.saltForUsername(blockchainIdentity.currentUsername!!, false)
-                    usernameStatus = blockchainIdentity.statusOfUsername(blockchainIdentity.currentUsername!!)
-                }
-                val requestStatus = blockchainIdentity.getUsernameRequestStatus(username!!)
-                if (requestStatus != UsernameRequestStatus.NONE) {
-                    usernameRequested = requestStatus
-                }
-                val votingStart = blockchainIdentity.getUsernameVotingStart(username!!)
-                if (votingStart != -1L) {
-                    votingPeriodStart = votingStart
-                }
-
-                log.info("creation: blockchainIdentity.secondaryUsername = {}", blockchainIdentity.secondaryUsername)
-                blockchainIdentity.secondaryUsername?.let { name ->
-                    usernameSecondary = name
-                    usernameSecondaryStatus = blockchainIdentity.statusOfUsername(name)
-                    log.info("creation: secondary username: {}, usernameSecondaryStatus = {}", name, usernameSecondaryStatus)
-                    preorderSaltSecondary = blockchainIdentity.saltForUsername(name, false)
-                }
-            }
-            creditBalance = blockchainIdentity.creditBalance
-
-        }
-        updateBlockchainIdentityData(blockchainIdentityData)
-    }
-
-    suspend fun resetIdentityCreationStateError(blockchainIdentityData: BlockchainIdentityData) {
-        blockchainIdentityDataStorage.updateCreationState(blockchainIdentityData.creationState, null)
-        blockchainIdentityData.creationStateErrorMessage = null
-    }
-
-    suspend fun updateIdentityCreationState(blockchainIdentityData: BlockchainIdentityData,
-                                            state: IdentityCreationState,
-                                            exception: Throwable? = null) {
-        val errorMessage = exception?.run {
-            var message = "${exception.javaClass.simpleName}: ${exception.message}"
-            if (this is StatusRuntimeException) {
-                val exceptionInfo = GrpcExceptionInfo(this).exception
-                message += exceptionInfo
-            }
-            message
-        }
-        if (errorMessage == null) {
-            log.info("updating creation state {}", state)
-        } else {
-            log.info("updating creation state {} ({})", state, errorMessage)
-        }
-        blockchainIdentityDataStorage.updateCreationState(state, errorMessage)
-        blockchainIdentityData.creationState = state
-        blockchainIdentityData.creationStateErrorMessage = errorMessage
-    }
-
-    suspend fun updateBlockchainIdentityData(blockchainIdentityData: BlockchainIdentityData) {
-        blockchainIdentityDataStorage.insert(blockchainIdentityData)
+    suspend fun updateDashPayProfile(dashPayProfile: DashPayProfile) {
+        dashPayProfileDao.insert(dashPayProfile)
     }
 
     /**
@@ -974,60 +373,12 @@ class PlatformRepo @Inject constructor(
         }
     }
 
-    suspend fun updateDashPayProfile(dashPayProfile: DashPayProfile) {
-        dashPayProfileDao.insert(dashPayProfile)
-    }
-
-    suspend fun updateDashPayContactRequest(dashPayContactRequest: DashPayContactRequest) {
-        dashPayContactRequestDao.insert(dashPayContactRequest)
-    }
-
-    suspend fun doneAndDismiss() {
-        val blockchainIdentityData = blockchainIdentityDataStorage.load()
-        if (blockchainIdentityData != null && blockchainIdentityData.creationState == IdentityCreationState.DONE) {
-            blockchainIdentityData.creationState = IdentityCreationState.DONE_AND_DISMISS
-            blockchainIdentityDataStorage.insert(blockchainIdentityData)
-        }
-    }
-
     //
     // Step 5: Find the usernames in the case of recovery
     //
     fun recoverUsernames(blockchainIdentity: BlockchainIdentity) {
         blockchainIdentity.recoverUsernames()
     }
-
-    //Step 6: Recover the DashPay Profile
-    suspend fun recoverDashPayProfile(blockchainIdentity: BlockchainIdentity) {
-        withContext(Dispatchers.IO) {
-            if (platform.hasApp("dashpay")) {
-                getUsername()?.let { username ->
-                    // recovery will only get the information and place it in the database
-                    val profile = blockchainIdentity.getProfile()
-
-
-                    // blockchainIdentity doesn't yet keep track of the profile, so we will load it
-                    // into the database directly
-                    val dashPayProfile = if (profile != null)
-                        DashPayProfile.fromDocument(profile, username)
-                    else
-                        DashPayProfile(blockchainIdentity.uniqueIdString, username)
-                    updateDashPayProfile(dashPayProfile)
-                }
-            }
-        }
-    }
-
-    fun getNextContactAddress(userId: String, accountReference: Int): Address? {
-        return try {
-            blockchainIdentity.getContactNextPaymentAddress(Identifier.from(userId), accountReference)
-        } catch (e: NullPointerException) {
-            log.error("Failed to get contact address due to null key chain", e)
-            null
-        }
-    }
-
-    var counterForReport = 0
 
     fun addSeriousErrorListener(listener: SeriousErrorListener) {
         onSeriousErrorListeneners.add(listener)
@@ -1052,11 +403,6 @@ class PlatformRepo @Inject constructor(
         return nameDocument.dashUniqueIdentityId ?: nameDocument.dashAliasIdentityId!!
     }
 
-    suspend fun getLocalUserProfile(): DashPayProfile? {
-        val blockchainIdentityBaseData = loadBlockchainIdentityBaseData()!!
-        return dashPayProfileDao.loadByUserId(blockchainIdentityBaseData.userId!!)
-    }
-
     suspend fun getLocalUserDataByUsername(username: String): UsernameSearchResult? {
         log.info("requesting local user data for $username")
         val profile = dashPayProfileDao.loadByUsername(username)
@@ -1078,24 +424,6 @@ class PlatformRepo @Inject constructor(
         }
     }
 
-
-
-    /**
-    This is used by java code, outside of coroutines
-
-    This should not be a suspended method.
-     */
-    suspend fun clearDatabase(includeInvitations: Boolean) {
-        log.info("clearing databases (includeInvitations = $includeInvitations)")
-        dashPayProfileDao.clear()
-        dashPayContactRequestDao.clear()
-        userAlertDao.clear()
-        clearBlockchainIdentityData()
-        if (includeInvitations) {
-            invitationsDao.clear()
-        }
-    }
-
     fun getBlockchainIdentityKey(index: Int, keyParameter: KeyParameter?): IDeterministicKey? {
         val authenticationChain = authenticationGroupExtension?.getKeyChain(
             AuthenticationKeyChain.KeyChainType.BLOCKCHAIN_IDENTITY
@@ -1110,19 +438,6 @@ class PlatformRepo @Inject constructor(
         val key = decryptedChain.getKey(index)
         Preconditions.checkState(key.path.last().isHardened)
         return key
-    }
-
-    fun getIdentityFromPublicKeyId(): Identity? {
-        return try {
-            getWalletEncryptionKey()?.let {
-                val firstIdentityKey = getBlockchainIdentityKey(0, it) ?: return null
-                platform.stateRepository.fetchIdentityFromPubKeyHash(firstIdentityKey.pubKeyHash)
-            }
-        } catch (e: MaxRetriesReachedException) {
-            null
-        } catch (e: NoAvailableAddressesForRetryException) {
-            null
-        }
     }
 
     fun observeProfileByUserId(userId: String): Flow<DashPayProfile?> {
@@ -1206,67 +521,9 @@ class PlatformRepo @Inject constructor(
         }
     }
 
-    suspend fun clearBlockchainIdentityData() {
-        blockchainIdentityDataStorage.clear()
-    }
-
-    // current unused
-    private suspend fun getContactRequestReport(): String {
-        val report = StringBuilder()
-        val profiles = dashPayProfileDao.loadAll()
-        val profilesById = profiles.associateBy({ it.userId }, { it })
-        report.append("Contact Requests (Sent) -----------------\n")
-        dashPayContactRequestDao.loadToOthers(blockchainIdentity.uniqueIdString).forEach {
-            val fromProfile = profilesById[it.userId]
-            report.append(it.userId)
-            if (fromProfile != null) {
-                report.append("(").append(fromProfile.username).append(")")
-            }
-            report.append(" -> ").append(it.toUserId)
-            val toProfile = profilesById[it.toUserId]
-            if (toProfile != null) {
-                report.append("(").append(toProfile.username).append(")")
-            }
-            report.append("\n")
-        }
-        report.append("Contact Requests (Received) -----------------\n")
-        dashPayContactRequestDao.loadFromOthers(blockchainIdentity.uniqueIdString).forEach {
-            val fromProfile = profilesById[it.userId]
-            report.append(it.userId)
-            if (fromProfile != null) {
-                report.append("(").append(fromProfile).append(")")
-            }
-            report.append(" -> ").append(it.toUserId)
-            val toProfile = profilesById[it.toUserId]
-            if (toProfile != null) {
-                report.append("(").append(toProfile.username).append(")")
-            }
-            report.append("\n")
-        }
-        return report.toString()
-    }
-
-    suspend fun getIdentityBalance(): CreditBalanceInfo? {
-        return withContext(Dispatchers.IO) {
-            try {
-                CreditBalanceInfo(platform.client.getIdentityBalance(blockchainIdentity.uniqueIdentifier))
-            } catch (e: Exception) {
-                log.error("Failed to get identity balance", e)
-                null
-            }
-        }
-    }
-
     suspend fun getIdentityBalance(identifier: Identifier): CreditBalanceInfo {
         return withContext(Dispatchers.IO) {
             CreditBalanceInfo(platform.client.getIdentityBalance(identifier))
         }
-    }
-
-    suspend fun addInviteUserAlert() {
-        // this alert will be shown or not based on the current balance and will be
-        // managed by NotificationsLiveData
-        val userAlert = UserAlert(UserAlert.INVITATION_NOTIFICATION_TEXT, UserAlert.INVITATION_NOTIFICATION_ICON)
-        userAlertDao.insert(userAlert)
     }
 }
